ToastCard to own component
