+++ frontend/src/components/ToastCard/ToastCard.module.scss
@@ -1,0 +1,71 @@
+$success-color: #4cd16d;
+$error-color: red;
+$info-color: #2b89e9;
+$warning-color: #ffbc22;
+
+.toastCard {
+  position: relative;
+  width: 500px;
+  padding: 16px;
+  padding-right: 32px;
+  background: white;
+  border: solid 1px #e0e7ff;
+
+  &.success {
+    border-left: 4px solid $success-color;
+    svg {
+      color: $success-color;
+    }
+  }
+  &.error {
+    border-left: 4px solid $error-color;
+    svg {
+      color: $error-color;
+    }
+  }
+  &.info {
+    border-left: 4px solid $info-color;
+    svg {
+      color: $info-color;
+    }
+  }
+  &.warning {
+    border-left: 4px solid $warning-color;
+    svg {
+      color: $warning-color;
+    }
+  }
+}
+
+.icon {
+  flex-shrink: 0;
+}
+
+.title {
+  display: flex;
+  align-items: center;
+  font-weight: 600;
+  gap: 6px;
+  font-size: 14px;
+}
+
+.message {
+  font-size: 12px;
+}
+
+.close {
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 16px;
+  cursor: pointer;
+
+  svg {
+    font-size: 12px;
+    color: gray !important;
+  }
+
+  &:hover svg {
+    transform: scale(1.3);
+  }
+}
